refactor(storage): inline acl entrypoints
